+++ ghost/core/core/server/services/lib/EmailContentGenerator.js
