+++ ghost/core/package.json
@@ -79,7 +79,6 @@
     "@tryghost/domain-events": "0.0.0",
     "@tryghost/donations": "0.0.0",
     "@tryghost/email-addresses": "0.0.0",
-    "@tryghost/email-content-generator": "0.0.0",
     "@tryghost/email-mock-receiver": "0.3.8",
     "@tryghost/email-service": "0.0.0",
     "@tryghost/email-suppression-list": "0.0.0",
@@ -169,6 +168,7 @@
     "gscan": "4.47.0",
     "handlebars": "4.7.8",
     "hcaptcha": "0.2.0",
+    "html-to-text": "5.1.1",
     "human-number": "2.0.4",
     "iconv-lite": "0.6.3",
     "image-size": "1.2.0",
