+++ ghost/core/test/unit/server/services/lib/fixtures/templates/newsletter.html
