+++ ghost/core/test/unit/server/services/lib/fixtures/templates/welcome.html
