--- ghost/email-content-generator/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/EmailContentGenerator');
